rename files + names for NEW condition LUT
