+++ domain-cc/cc-app/src/python_src/api.py
@@ -6,10 +6,10 @@ from fastapi import FastAPI, HTTPException
 
 from .pydantic_models import Claim, PredictedClassification
 from .util.brd_classification_codes import get_classification_name
-from .util.lookup_table import DiagnosticCodeLookupTable, DropdownLookupTable
+from .util.lookup_table import DiagnosticCodeLookupTable, ConditionDropdownLookupTable
 
 dc_lookup_table = DiagnosticCodeLookupTable()
-dropdown_lookup_table = DropdownLookupTable()
+dropdown_lookup_table = ConditionDropdownLookupTable()
 
 
 app = FastAPI(
