+++ domain-cc/cc-app/src/python_src/util/condition_dropdown_table_version.py
@@ -1,0 +1,1 @@
+CONDITION_DROPDOWN_TABLE_VERSION = "v0.1"
