+++ domain-cc/cc-app/src/python_src/util/diagnostic_code_table_version.py
