--- domain-cc/cc-app/src/python_src/util/dropdown_table_version.py
@@ -1,1 +1,0 @@
-DROPDOWN_TABLE_VERSION = "v0.1"
